[#943] Remove pinia logout plugin
